+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Improve browser compatibility
+
 ## [1.0.3] 2025-02-16
 
 ### Changed