+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Fixed
 
 - Improve browser compatibility
+- Fix mod menu sometimes opening empty
+- Improve cc-instanceinator compatibility
 
 ## [1.0.3] 2025-02-16