+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # Change Log
 
 ## [Unreleased]
+## [1.0.4] 2025-06-26
 
 ### Added